requirements.txt
# FastAPI and web framework
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
pydantic[email]>=2.5.0,<3.0.0

# Database
pymongo>=4.5,<5.0
motor>=3.3.0,<4.0.0
dnspython>=2.3.0

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
google-auth>=2.20.0

# AWS S3
boto3>=1.28.0
botocore>=1.31.0

# HTTP client
httpx>=0.25.0
requests>=2.31.0

# Configuration
python-decouple>=3.8

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0